Doing something wrong, off by a factor of two
So we expect the average force applied to the surface to be equal to
(well, close to since we have discrete rounding errors) the pressure,
which we expect to be equal to the total kinetic energy since

PV = nRT

unless R is almost exactly 2 for some reason. So what do we expect?

Well the force applied during a collision is 2*p/dt (one p to stop it,
another to make it -p, and then divide by the time this took). So the
total force is going to be (2*p/dt)*collisions*dt. We expect the number
of collisions to go like (N/T), where T is the time to go from one
end to another, and this will scale with N and p. So an individual
particle will contribute be 2*p**2. So since the total kinetic
energy is p**2/2, so we're off by a factor of four for some reason.
Actually, this is fine, if we're only looking at one side of the
chamber, we expect to travel twice the distance, so we pick up another
factor of two. We're still off by a factor of two though.