The box can now be expanded!
As part of the ongoing process of allowing us to simulate PV=nRT
behavior, I have added the ability to control the size of the box.

Of note, expanding the box will eventually saturate the average force,
mostly because you get the initial collisions on the left side, which
count for a lot. To correct this, you can run the simulation time for
longer.

Science!